+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "moocadmin",
-  "version": "1.0.33",
+  "version": "1.0.35",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.14.0",
